+++ CHANGELOG.md
@@ -5,11 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.6.2] - Unreleased
+## [0.6.2] - cobra-extensions v0.6.2, 2026-07-17
 
 ### Added
 
 * Support for mapping flags specified multiple times to slice fields in command structs (supported types: `[]string`, `[]int`, `[]int64`, `[]bool`).
+* Support for short flag names (shorthands) via the `shorthand` struct tag.
 
 ### Changed